Revert "Better infractions"
