+++ src/commands/moderation/ban.js
@@ -1,7 +1,6 @@
+const Emotes = require("../../emotes.json");
 const Log = require("../../utils/moderation/log");
 const Moderation = require("../../utils/moderation/moderation");
-const Translator = require("../../utils/lang/translator")
-const Global = require("../../utils/database/global")
 
 module.exports = {
 	name: "ban",
@@ -20,12 +19,11 @@ module.exports = {
 	run: async (client, message, args) => {
 		if (args[0] === undefined || args[0] === null)
 			return message.channel.send(
-				Translator.Translate("ban_missing_arg_user")
+				`${Emotes.actions.warn} Missing required argument \`\`user\`\`\n${Emotes.other.tools} Correct usage of command: \`\`ban|terminate|yeet <user> [reason]\`\``
 			);
 		let target = args[0].replace(/\D/g, ""); // Remove everything except numbers
 		let user = message.guild.member(target);
 		let reason = args.slice(1).join(" ");
-		const infID = await Global.NumberInfraction();
 
 		const banList = await message.guild.fetchBans();
 		const bannedUser = banList.find((user) => user.user.id === target);
@@ -34,7 +32,7 @@ module.exports = {
 			return message.channel.send(
 				`<@${target}> \`\`(${target})\`\` is already banned.`
 			);
-    
+
 		if (reason === "") reason = "No reason given";
 
 		if (user === null) {
@@ -45,8 +43,8 @@ module.exports = {
 			}
 
 			if (user === null)
-				return message.channel.send(
-					Translator.Translate("global_user_not_found", {user: target})
+				return await message.channel.send(
+					"Unable to find the user, please try again with the correct **user id**"
 				);
 			if (
 				!(await Moderation.ForceBan(
@@ -58,11 +56,7 @@ module.exports = {
 				))
 			)
 				return message.channel.send(
-					Translator.Translate("ban_fail", {
-						user: user.user.username,
-						user_discriminator: user.user.discriminator,
-						user_id: user.user.id
-					})
+					`Unable to ban <@${target}> \`\`(${target})\`\`.`
 				);
 		} else {
 			if (
@@ -75,12 +69,7 @@ module.exports = {
 				))
 			)
 				return message.channel.send(
-					Translator.Translate("ban_fail", {
-						user: user.user.username,
-						user_discriminator: user.user.discriminator,
-						user_id: user.user.id,
-						inf_number: infID
-					})
+					`Unable to ban <@${target}> \`\`(${target})\`\`.`
 				);
 			user = user.user;
 		}
@@ -88,27 +77,12 @@ module.exports = {
 		await Log.Mod_action(
 			client,
 			message.guild.id,
-			Translator.Translate("ban_log", {
-				user: user.username,
-				user_discriminator: user.discriminator,
-				user_id: user.id,
-				moderator: message.author.username,
-				moderator_discriminator: message.author.discriminator,
-				moderator_id: message.author.id,
-				reason: reason,
-				inf_number: infID
-			}),
+			`${Emotes.actions.ban} Banned **${user.username}**#${user.discriminator} \`\`(${user.id})\`\` by **${message.author.username}**#${message.author.discriminator} \`\`(${message.author.id})\`\` \n**Reason:** ${reason} `,
 			""
 		);
 
 		message.channel.send(
-			Translator.Translate("ban_success", {
-				user: user.username,
-				user_discriminator: user.discriminator,
-				user_id: user.id,
-				reason: reason,
-				inf_number: infID
-			})
+			`${Emotes.actions.ban} Banning <@${target}> \`\`(${target})\`\` for \`\`${reason}\`\``
 		);
 	},
 };
