+++ src/commands/moderation/kick.js
@@ -1,7 +1,6 @@
 const Moderation = require("../../utils/moderation/moderation");
 const Log = require("../../utils/moderation/log");
-const Translator = require("../../utils/lang/translator")
-const Global = require("../../utils/database/global")
+const Emotes = require("../../emotes.json");
 
 module.exports = {
 	name: "kick",
@@ -19,14 +18,13 @@ module.exports = {
 	run: async (client, message, args) => {
 		if (args[0] === undefined || args[0] === null)
 			return message.channel.send(
-				Translator.Translate("kick_missing_arg_user")
+				`${Emotes.actions.warn} Missing required argument \`\`user\`\`\n${Emotes.other.tools} Correct usage of command: \`\`kick <user> [reason]\`\``
 			);
 		let target = args[0].replace(/\D/g, ""); // Remove everything except numbers
 		let user = message.guild.member(target);
-		let reason = args.slice(1).join(" ")
-		const infID = await Global.NumberInfraction();
+		let reason = args.slice(1).join(" ");
 		if (reason === "") reason = "No reason given";
-		if (user === null) return message.channel.send(Translator.Translate("global_user_not_found"));
+		if (user === null) return message.channel.send("User is not in guild.");
 
 		if (
 			!(await Moderation.Kick(
@@ -38,36 +36,17 @@ module.exports = {
 			))
 		)
 			return message.channel.send(
-				Translator.Translate("kick_fail", {
-					user: user.user.username,
-					user_discriminator: user.user.discriminator,
-					user_id: user.user.id
-				})
+				`Unable to kick <@${target}> \`\`(${target})\`\`.`
 			);
 		await Log.Mod_action(
 			client,
 			message.guild.id,
-			Translator.Translate("kick_log", {
-				user: user.user.username,
-				user_discriminator: user.user.discriminator,
-				user_id: user.user.id,
-				moderator: message.author.username,
-				moderator_discriminator: message.author.discriminator,
-				moderator_id: message.author.id,
-				reason: reason,
-				inf_number: infID
-			}),
+			`${Emotes.actions.kick} Kicked **${user.user.username}**#${user.user.discriminator} \`\`(${user.user.id})\`\` by **${message.author.username}**#${message.author.discriminator} \`\`(${message.author.id})\`\` \n**Reason:** ${reason} `,
 			""
 		);
 
 		message.channel.send(
-			Translator.Translate("kick_success", {
-				user: user.user.username,
-				user_discriminator: user.user.discriminator,
-				user_id: user.user.id,
-				reason: reason,
-				inf_number: infID
-			})
+			`${Emotes.actions.kick} Kicking <@${target}> \`\`(${target})\`\` for \`\`${reason}\`\``
 		);
 	},
 };
