+++ src/commands/moderation/multiban.js
@@ -1,6 +1,6 @@
 const Moderation = require("../../utils/moderation/moderation");
 const Log = require("../../utils/moderation/log");
-const Translator = require("../../utils/lang/translator")
+const Emotes = require("../../emotes.json");
 
 module.exports = {
 	name: "multiban",
@@ -14,7 +14,7 @@ module.exports = {
 	run: async (client, message, args) => {
 		if (args[0] === undefined || args[0] === null)
 			return message.channel.send(
-				Translator.Translate("mban_missing_arg_users")
+				`${Emotes.actions.warn} Missing required argument \`\`users\`\`\n${Emotes.other.tools} Correct usage of command: \`\`multiban <user> [user2]...[reason]\`\``
 			);
 
 		const targets = args
@@ -41,31 +41,18 @@ module.exports = {
 					reason
 				))
 			)
-				message.channel.send(Translator.Translate("ban_fail", {
-					user: t.user.username,
-					user_discriminator: t.user.discriminator,
-					user_id: t.user.id
-				}));
+				message.channel.send(`Unable to ban <@${target}> \`\`(${target})\`\`.`);
 			fullList += `**${t.username}**#${t.discriminator} \`\`(${t.id})\`\`, `;
 		}
 
 		await Log.Mod_action(
 			client,
 			message.guild.id,
-			Translator.Translate("mkick_log", {
-				new_value: fullList,
-				reason: reason,
-				moderator: message.author.username,
-				moderator_discriminator: message.author.discriminator,
-				moderator_id: message.author.id,
-			}),
+			`${Emotes.actions.ban} Multiple targets were banned ${fullList} by **${message.author.username}**#${message.author.discriminator} \`\`(${message.author.id})\`\` \n**Reason:** ${reason} `,
 			""
 		);
 		message.channel.send(
-			Translator.Translate("mkick_success", {
-				new_value: fullList,
-				reason: reason
-			})
+			`${Emotes.actions.ban} Banned ${fullList} for \`\`${reason}\`\``
 		);
 	},
 };
