+++ src/commands/moderation/softban.js
@@ -1,7 +1,6 @@
 const Moderation = require("../../utils/moderation/moderation");
 const Log = require("../../utils/moderation/log");
-const Global = require("../../utils/database/global")
-const Translator = require("../../utils/lang/translator")
+const Emotes = require("../../emotes.json");
 
 module.exports = {
 	name: "softban",
@@ -20,17 +19,14 @@ module.exports = {
 	run: async (client, message, args) => {
 		if (args[0] === undefined || args[0] === null)
 			return message.channel.send(
-				Translator.Translate("softban_missing_arg_user")
+				`${Emotes.actions.warn} Missing required argument \`\`user\`\`\n${Emotes.other.tools} Correct usage of command: \`\`softban|cleanban <user> [reason]\`\``
 			);
 		let target = args[0].replace(/\D/g, ""); // Remove everything except numbers
 		let user = message.guild.member(target);
 		let reason = args.slice(1).join(" ");
-		const infID = await Global.NumberInfraction();
 		if (reason === "") reason = "No reason given";
 		if (user === null) {
-			return message.channel.send(Translator.Translate("global_user_not_found", {
-				user: target
-			}));
+			return await message.channel.send("User is not in guild.");
 		} else {
 			if (
 				!(await Moderation.Softban(
@@ -42,11 +38,7 @@ module.exports = {
 				))
 			)
 				return message.channel.send(
-					Translator.Translate("softban_fail", {
-						user: user.user.username,
-						user_discriminator: user.user.discriminator,
-						user_id: user.user.id
-					})
+					`Unable to softban <@${target}> \`\`(${target})\`\`.`
 				);
 			user = user.user;
 		}
@@ -54,27 +46,12 @@ module.exports = {
 		await Log.Mod_action(
 			client,
 			message.guild.id,
-			Translator.Translate("softban_log", {
-				user: user.username,
-				user_discriminator: user.discriminator,
-				user_id: user.id,
-				moderator: message.author.username,
-				moderator_discriminator: message.author.discriminator,
-				moderator_id: message.author.id,
-				reason: reason,
-				inf_number: infID
-			}),
+			`${Emotes.actions.ban} Softbanned **${user.username}**#${user.discriminator} \`\`(${user.id})\`\` by **${message.author.username}**#${message.author.discriminator} \`\`(${message.author.id})\`\` \n**Reason:** ${reason} `,
 			""
 		);
 
 		message.channel.send(
-			Translator.Translate("softban_success", {
-				user: user.username,
-				user_discriminator: user.discriminator,
-				user_id: user.id,
-				reason: reason,
-				inf_number: infID
-			})
+			`${Emotes.actions.ban} Softbanning <@${target}> \`\`(${target})\`\` for \`\`${reason}\`\``
 		);
 	},
 };
