+++ src/commands/moderation/unban.js
@@ -1,8 +1,6 @@
 const Moderation = require("../../utils/moderation/moderation");
 const Log = require("../../utils/moderation/log");
 const Emotes = require("../../emotes.json");
-const Global = require("../../utils/database/global")
-const Translator = require("../../utils/lang/translator")
 
 module.exports = {
 	name: "unban",
@@ -22,11 +20,10 @@ module.exports = {
 	run: async (client, message, args) => {
 		if (args[0] === undefined || args[0] === null)
 			return message.channel.send(
-				Translator.Translate("unban_missing_arg_user")
+				`${Emotes.actions.warn} Missing required argument \`\`user\`\`\n${Emotes.other.tools} Correct usage of command: \`\`unban|pardon <user> [reason]\`\``
 			);
 		let target = args[0];
 		let reason = args.slice(1).join(" ");
-		const infID = await Global.NumberInfraction();
 		if (reason === "") reason = "No reason given";
 
 		let whoToUnban = "";
@@ -43,20 +40,14 @@ module.exports = {
 
 			if (whoToUnban === "")
 				return message.channel.send(
-					Translator.Translate("unban_fail", {
-						user: target
-					})
+					`Was unable to find user \`\`${target}\`\` in **${message.guild.name}**, try to be a bit more specific or check if they are banned.`
 				);
 
 			const user = await client.users.fetch(whoToUnban);
 
 			message.channel
 				.send(
-					Translator.Translate("unban_confirm", {
-						user: user.username,
-						user_discriminator: user.discriminator,
-						user_id: user.id
-					})
+					`Are you sure you want to unban **${user.username}**#${user.discriminator} \`\`(${user.id})\`\`?`
 				)
 				.then((msg) => {
 					msg
@@ -94,40 +85,21 @@ module.exports = {
 									))
 								)
 									return message.channel.send(
-										Translator.Translate("unban_fail_2", {
-											user: user.username,
-											user_discriminator: user.discriminator,
-											user_id: user.id
-										})
+										`Unable to unban <@${user.id}> \`\`(${user.id})\`\`.`
 									);
 								await Log.Mod_action(
 									client,
 									message.guild.id,
-									Translator.Translate("unban_log", {
-										user: user.username,
-										user_discriminator: user.discriminator,
-										user_id: user.id,
-										moderator: message.author.username,
-										moderator_discriminator: message.author.discriminator,
-										moderator_id: message.author.id,
-										reason: reason,
-										inf_number: infID
-									}),
+									`${Emotes.actions.unban} Unbanned **${user.username}**#${user.discriminator} \`\`(${user.id})\`\` by **${message.author.username}**#${message.author.discriminator} \`\`(${message.author.id})\`\` \n**Reason:** ${reason} `,
 									""
 								);
 
 								return message.channel.send(
-									Translator.Translate("unban_success", {
-										user: user.username,
-										user_discriminator: user.discriminator,
-										user_id: user.id,
-										reason: reason,
-										inf_number: infID
-									})
+									`${Emotes.actions.unban} Unbanning <@${whoToUnban}> \`\`(${whoToUnban})\`\` for \`\`${reason}\`\``
 								);
 							} else {
 								return message.channel.send(
-									Translator.Translate("global_operation_cancel")
+									`${Emotes.actions.cancel} Canceling the operation.`
 								);
 							}
 						});
