+++ src/commands/moderation/warn.js
@@ -1,7 +1,6 @@
 const Moderation = require("../../utils/moderation/moderation");
 const Log = require("../../utils/moderation/log");
-const Translator = require("../../utils/lang/translator")
-const Global = require("../../utils/database/global")
+const Emotes = require("../../emotes.json");
 
 module.exports = {
 	name: "warn",
@@ -14,17 +13,13 @@ module.exports = {
 	run: async (client, message, args) => {
 		if (args[0] === undefined || args[0] === null)
 			return message.channel.send(
-				Translator.Translate("warn_missing_arg_user")
+				`${Emotes.actions.warn} Missing required argument \`\`user\`\`\n${Emotes.other.tools} Correct usage of command: \`\`warn <user> [reason]\`\``
 			);
 		let target = args[0].replace(/\D/g, ""); // Remove everything except numbers
 		let user = message.guild.member(target);
 		let reason = args.slice(1).join(" ");
-		let inf_ID = await Global.NumberInfraction();
 		if (reason === "") reason = "No reason given";
-		if (user === null) return message.channel.send(
-			Translator.Translate("global_user_not_found", {
-				user: target
-			}));
+		if (user === null) return message.channel.send("User is not in server");
 
 		await Moderation.Warn(
 			client,
@@ -37,27 +32,12 @@ module.exports = {
 		await Log.Mod_action(
 			client,
 			message.guild.id,
-			Translator.Translate("warn_log", {
-				user: user.user.username,
-				user_discriminator: user.user.discriminator,
-				user_id: user.user.id,
-				moderator: message.author.username,
-				moderator_discriminator: message.author.discriminator,
-				moderator_id: message.author.id,
-				reason: reason,
-				inf_number: inf_ID
-			}),
+			`${Emotes.actions.warn} Warning given to **${user.user.username}**#${user.user.discriminator} \`\`(${user.user.id})\`\` by **${message.author.username}**#${message.author.discriminator} \`\`(${message.author.id})\`\` \n**Reason:** ${reason} `,
 			""
 		);
 
 		message.channel.send(
-			Translator.Translate("warn_success", {
-				user: user.user.username,
-				user_discriminator: user.user.discriminator,
-				user_id: user.user.id,
-				reason: reason,
-				inf_number: inf_ID
-			})
+			`${Emotes.actions.warn} Warned <@${target}> \`\`(${target})\`\` for \`\`${reason}\`\``
 		);
 	},
 };
