+++ src/events/guildBanRemove.js
@@ -1,11 +1,7 @@
 const Log = require("../utils/moderation/log");
 const Translator = require("../utils/lang/translator")
-const Moderation = require("../utils/moderation/moderation")
-const Global = require("../utils/database/global")
 
 module.exports = async (client, guild, user) => {
-    const inf_ID = await Global.NumberInfraction();
-
     const log = await guild.fetchAuditLogs({limit: 1, type: 'MEMBER_BAN_REMOVE'});
     const banLog = log.entries.first();
     const { executor } = banLog;
@@ -22,10 +18,7 @@ module.exports = async (client, guild, user) => {
                 moderator: executor.username,
                 moderator_discriminator: executor.discriminator,
                 moderator_id: executor.id,
-                inf_number: inf_ID
             }),
         ""
     )
-
-    await Moderation.Unban(client, guild.id, user.id, executor.id, "Manual Unban")
 }
