--- src/models/global.js
@@ -1,9 +1,0 @@
-const mongoose = require("mongoose");
-
-const globalSchema = mongoose.Schema({
-	infCounter: Number,
-	commandsFire: Number,
-	totalMessages: Number,
-});
-
-module.exports = mongoose.model("Global", globalSchema);
