+++ src/models/infraction.js
@@ -3,7 +3,6 @@ const mongoose = require("mongoose");
 const infractionSchema = mongoose.Schema({
   guildID: String,
   action: String,
-  infID: Number,
   targetID: String,
   moderatorID: String,
   reportReason: String,
