+++ src/utils/lang/translator.js
@@ -24,14 +24,10 @@ module.exports = {
 			time,
 			nick_old,
 			nick_new,
-			channel_id,
-			inf_number
+			channel_id
 		} = {}
 	) => {
 		let response = JSON.parse(JSON.stringify(lang))[key].toString();
-		//Infractions
-		response = response.replace(/({inf_number})/g, inf_number);
-
 		//Uptime
 		response = response.replace(/({uptime})/g, uptime);
